third_party/blink/web_tests/css3/calc/getComputedStyle-border-radius-expected.txt

Tests calling getPropertyValue on computed border radii involving CSS calc

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(document.getElementById('calc'), null).getPropertyValue('border-top-left-radius') is "calc(25% + 10px) calc(25% + 20px)"
PASS getComputedStyle(document.getElementById('calc'), null).getPropertyValue('border-top-right-radius') is "calc(25% + 10px)"
PASS getComputedStyle(document.getElementById('calc'), null).getPropertyValue('border-bottom-right-radius') is "25%"
PASS getComputedStyle(document.getElementById('calc'), null).getPropertyValue('border-bottom-left-radius') is "25px"
PASS getComputedStyle(document.getElementById('calc'), null).getPropertyValue('border-radius') is "calc(25% + 10px) calc(25% + 10px) 25% 25px / calc(25% + 20px) calc(25% + 10px) 25% 25px"
PASS successfullyParsed is true

TEST COMPLETE
